Set gdb/version.in to 7.6.2.20131208-cvs.
This replaces <DATE> with today's actual date in the format expected
by the branch nightly update script.  The use of the <DATE> placeholder
will be appropriate for 7.7, where it is automatically replace during
the build from BFD's version.h. But for 7.6, it should be something
that looks like a date.

gdb/ChangeLog:

	* version.in: Set GDB version number to 7.6.2.20131208-cvs.